requirements.txt
South
django-mptt>=0.5.2
six
unicode-slugify>=0.1.1